refactor: formatDate -> format
