+++ app/src/lib/format.ts
@@ -1,6 +1,6 @@
 import format from 'date-fns/format';
 
-const formatDate = (date: Date | string) =>
+export const formatDate = (date: Date | string) =>
   format(new Date(date), 'dd MMMM yyyy');
 
 export const safeFormatDate = (date: Date | string) => {
@@ -10,5 +10,3 @@ export const safeFormatDate = (date: Date | string) => {
 };
 
 export const datePickerFormat = (date: Date) => format(date, 'yyyy-MM-dd');
-
-export default formatDate;
